backend/pyproject.toml
[tool.poetry]
name = "voice-clone-chat"
version = "0.1.0"
description = "Chat with an AI version of anyone"
authors = ["Sam Liang <iamsamliang8@gmail.com>"]
license = "MIT"
readme = "../README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "~3.11.6"
fastapi = "^0.104.1"
openai = "^1.3.5"
deepgram-sdk = "^2.12.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
uvicorn = "^0.24.0.post1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"
ipykernel = "^6.26.0"
types-requests = "^2.31.0.10"
pydub = "^0.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
